Moved P2P init into one function.
